rename PropertyNameReplacerRector to RenamePropertyRector
