+++ config/level/laravel/laravel55.yaml
@@ -4,7 +4,7 @@ services:
         Illuminate\Console\Command:
             fire: 'handle'
 
-    Rector\Rector\Property\PropertyNameReplacerRector:
+    Rector\Rector\Property\RenamePropertyRector:
         Illuminate\Database\Eloquent\Concerns\HasEvents:
             events: 'dispatchesEvents'
         Illuminate\Database\Eloquent\Relations\Pivot:
