--- tests/Rector/Property/PropertyNameReplacerRector/Source/ClassWithProperties.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\Property\PropertyNameReplacerRector\Source;
-
-final class ClassWithProperties
-{
-
-}
