+++ tests/Rector/Property/RenamePropertyRector/Source/ClassWithProperties.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\Property\RenamePropertyRector\Source;
+
+final class ClassWithProperties
+{
+
+}
